backend/requirements.txt
fastapi
uvicorn
python-multipart
pyyaml
pillow
dicttoxml
beautifulsoup4
xhtml2pdf==0.2.15
markdownify==0.11.6
html2image==2.0.4.3
python-docx
pymupdf
pdf2docx
